requirements.txt
google-cloud-documentai==2.31.0
google-cloud-storage==2.12.0
google-cloud-aiplatform==1.54.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
jupyterlab==4.0.8
python-dotenv==1.0.0
# Fixed versions for compatibility
huggingface_hub==0.19.4
transformers==4.35.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
openai==1.3.6
weaviate-client==3.25.3
streamlit==1.28.2
Pillow==10.1.0
PyPDF2==3.0.1
torch>=1.13.0
